refactor: move load_external to home/previews
