+++ src/pages/home/previews/xls.tsx
@@ -1,6 +1,6 @@
 import { BoxWithFullScreen, Error as Erro, FullLoading } from "~/components"
 import { Box, IconButton, Tooltip, Button, HStack } from "@hope-ui/solid"
-import { loadScript } from "~/utils"
+import { loadScript } from "./load_external"
 import { createSignal, onMount, For, Show } from "solid-js"
 import { useLink, useT } from "~/hooks"
 import { VsScreenFull, VsScreenNormal } from "solid-icons/vs"
