+++ src/utils/index.ts
@@ -13,4 +13,3 @@ export * from "./hash"
 export * from "./compatibility"
 export * from "./share"
 export * from "./storage"
-export * from "./asset"
